feat: add prisma 7 support (#1)

+++ test/scripts/run-with-postgres.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 
-export DATABASE_URL=postgres://postgres:123@localhost:5433/test
+export DATABASE_URL=postgres://postgres:postgrespass@localhost:5433/test
 
-trap "docker compose down" EXIT
-
-docker compose up -d && sleep 1
+if [ "$1" != "ci" ]; then
+  trap "docker compose down" EXIT
+  docker compose up -d && sleep 1
+else
+  shift
+fi
 npx prisma db push
 
 $@
